+++ src/lib.rs
@@ -1,10 +1,14 @@
 use anyhow::Result;
 
+mod component;
+mod game;
+mod math;
+
 pub const VERSION: &'static str = env!("CARGO_PKG_VERSION");
 pub const GIT_HASH: &'static str = env!("GIT_HASH");
 
 pub fn run() -> Result<()> {
     println!("spellthief v{} [{}]", VERSION, GIT_HASH);
 
-    Ok(())
+    game::main_loop()
 }