+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### added
 
+- `hecs` crate to manage game and entity state
+
 ### changed
 
 ### fixed